Added license, used `write_status`, removed 'Config.cmake.in'
Adding a brief and license text at the top of the file as comments.
Using `write_status` function instead of normal `message(STATUS...` function.
Using KDI namespace variables instead of the double underscore ones.
Removing 'Config.cmake.in' and using 'config-shared.cmake.in' for shared library and 'config-static.cmake.in' for static library.